docker/backend/requirements.txt
flask==2.2.2
flask-cors==3.0.10
werkzeug==2.2.2
pillow==9.5.0
mysql-connector-python==8.0.33
numpy==1.24.3
pandas==2.0.2
opencv-python==4.7.0.72
mediapipe==0.10.8
scikit-learn==1.2.2
